Extract part of DB to new SchemaTypeChecking trait.
